backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
bcrypt
passlib[bcrypt]
email-validator
pydantic
aiosmtplib
PyJWT